Fix issue #5. For PG >= 16, don't call pg_sequence_parameters().
The implementation of pg_sequence_parameters() and some other
set-returning functions in core has been changed in v16 to get
the tuple structure from fcinfo.
That results in a crash when calling it through
DirectFunctionCall1(pg_sequence_parameters, seq_oid)
as we did before.
Therefore for v16+, this code now gets at the sequence tuple with
SearchSysCache1(SEQRELID, seq_oid).